create badge component
