+++ src/domain/orders/details/swap/timeline.js
@@ -8,7 +8,7 @@ import ReactTooltip from "react-tooltip"
 import { ReactComponent as Clipboard } from "../../../../assets/svg/clipboard.svg"
 import { decideBadgeColor } from "../../../../utils/decide-badge-color"
 import Typography from "../../../../components/typography"
-import Badge from "../../../../components/badge"
+import Badge from "../../../../components/fundamentals/badge"
 import LineItem from "../line-item"
 import Dropdown from "../../../../components/dropdown"
 import useMedusa from "../../../../hooks/use-medusa"
