+++ src/domain/orders/index.js
@@ -30,7 +30,7 @@ import {
   DefaultCellContent,
   BadgdeCellContent,
 } from "../../components/table"
-import Badge from "../../components/badge"
+import Badge from "../../components/fundamentals/badge"
 
 import { displayAmount } from "../../utils/prices"
 import { decideBadgeColor } from "../../utils/decide-badge-color"
